requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
cffi==1.2.1               # via cryptography
click==5.1
cryptography==1.0.2       # via pyopenssl
enum34==1.0.4             # via cryptography
idna==2.0                 # via cryptography
ipaddress==1.0.14         # via cryptography
ndg-httpsclient==0.4.0
oauthlib==1.0.3           # via requests-oauthlib
pyasn1==0.1.9
pycparser==2.14           # via cffi
pyopenssl==0.15.1
pyyaml==3.11
requests-oauthlib==0.5.0
requests==2.8.1
six==1.10.0               # via cryptography, pyopenssl

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools==18.4          # via cryptography
